requirements.txt
requests>=2.0
python-dotenv>=0.0.0
openai==2.21.0
streamlit==1.54.0
yfinance==1.1.0
tiktoken==0.12.0
langchain-openai==1.1.9
langgraph==1.0.8
langchain==1.2.10
langchain_community==0.4.1
chainlit==2.9.6